gyro io interface and code reorganization
