+++ src/main/java/frc/robot/RobotContainer.java
@@ -1,9 +1,9 @@
 package frc.robot;
 
-import frc.robot.subsystems.Drivetrain;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.RunCommand;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
+import frc.robot.subsystems.drivetrain.Drivetrain;
 
 public class RobotContainer {
 	// controllers
