+++ src/main/java/frc/robot/util/SwerveUtil.java
@@ -1,4 +1,4 @@
-package frc.robot.swerve;
+package frc.robot.util;
 
 import static frc.robot.constants.MathConstants.kTau;
